Rename backend and ensure module loading is more reliable
